requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
typing-extensions
numpy
requests
python-multipart
starlette
google-generativeai
tavily-python
pydantic-settings